requirements.txt
beautifulsoup4==4.12.3
GitPython==3.1.43
Markdown==3.7
PyYAML==6.0.2
loguru==0.7.2
